requirements.txt
pandas
plotly
streamlit
openpyxl
matplotlib
streamlit_option_menu
msoffcrypto-tool
streamlit_authenticator
xlrd
deta
streamlit_lottie
Pillow
seaborn
